JWT Auth added & header/footer

+++ client/app/api/notes/delete/[folder]/route.ts
@@ -1,23 +1,34 @@
 import prisma from '@/app/lib/prisma';
 import { NextRequest, NextResponse } from 'next/server';
-import { Post } from '../../all/route';
-
 
 export async function DELETE(req: NextRequest) {
   try {
+    const userId = Number(req.headers.get('x-user-id'));
     const { id } = await req.json();
+
+    const post = await prisma.posts.findFirst({
+      where: {
+        id: Number(id),
+        userId
+      }
+    });
+
+    if (!post) {
+      return NextResponse.json(
+        { error: 'Заметка не найдена или у вас нет прав на её удаление' },
+        { status: 404 }
+      );
+    }
     await prisma.posts.delete({
-        where: { id: Number(id) },
-      }) as Post;
-      return NextResponse.json(id)
-    } 
-  catch (error) {
+      where: { id: Number(id) }
+    });
+
+    return NextResponse.json(id);
+  } catch (error) {
     console.error('Ошибка при удалении заметки:', error);
     return NextResponse.json(
-      {
-        error: 'Ошибка при удалении заметки: ' + error,
-      },
-      { status: 500 },
+      { error: 'Ошибка при удалении заметки' },
+      { status: 500 }
     );
   }
 }
